requirements.txt
django-ckeditor
Pillow
pylint
pylint-django
pylint-celery
django-cors-headers
djangorestframework
djangorestframework-simplejwt
openpyxl